requirements.txt
fastapi 
uvicorn 
python-multipart 
tensorflow 
pillow 
jinja2
gradio
numpy
deep-translator
